refactor: 패키지 구조 정리 및 설정(config) 분리
