+++ catalog-service/src/main/java/com/node5/catalogservice/config/s3/S3PresignerConfig.java
@@ -1,4 +1,4 @@
-package com.node5.catalogservice.product.infrastructure.s3;
+package com.node5.catalogservice.config.s3;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
