+++ catalog-service/src/main/java/com/node5/catalogservice/config/search/SearchIndexProperties.java
@@ -1,4 +1,4 @@
-package com.node5.catalogservice.search.infrastructure.elasticsearch;
+package com.node5.catalogservice.config.search;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
