+++ catalog-service/src/main/java/com/node5/catalogservice/product/infrastructure/s3/S3PresignedUrlAdapter.java
@@ -5,6 +5,7 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Component;
 
+import com.node5.catalogservice.config.s3.S3Properties;
 import com.node5.catalogservice.product.application.dto.PresignedUrlInfo;
 import com.node5.catalogservice.product.application.port.S3PresignedUrlPort;
